refactor(auto-rejoin): moved auto-rejoin to automation folder and fixed missing function
